Added "pragmatic" syntax highlighting
Added alternate “pragmatic” syntax highlighting for sublime text 2.
“Pragmatic” highlighting does not reflect the underlying syntax or
semantics of the language, because it makes a meaningless distinction
between “keywords”, “built-in functions,” and “user-defined functions,”
all of which are dealt with by the interpreter in exactly the same way
and should be thought of as existing at an equal level.  Still, for
those who like to think in those terms, this color scheme highlights
Comet’s closest equivalents to those concepts.